move save metadata submenu to hook
